// src/enum/dom-manipulation-event.enum.ts
enum DOMManipulationEventEnum {
  MESSAGE_SENT = 'MESSAGE_SENT',
  UPDATED_PROMPT = 'UPDATED_PROMPT',
  CLEAN_PROMPT = 'CLEAN_PROMPT',
  ROLLED_DOWN = 'ROLLED_DOWN',
  STREAM_STARTED = 'STREAM_STARTED',
  STREAM_COMPLETED = 'STREAM_COMPLETED',
  AUDIO_PLAYING = 'AUDIO_PLAYING',
  AUDIO_STOPPED = 'AUDIO_STOPPED',
}

export default DOMManipulationEventEnum;